requirements.txt
# CoolBits.ai Requirements
# ========================

# Core web framework
Flask==3.0.0
FastAPI==0.104.1
uvicorn==0.24.0

# HTTP client
requests==2.31.0

# System monitoring
psutil==5.9.6

# CORS support
flask-cors==4.0.0

# Development dependencies
pytest==7.4.3
flake8==6.1.0
black==23.11.0

# Production WSGI server
gunicorn==21.2.0

# Environment management
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Security
cryptography==41.0.7

# Google Cloud (if needed)
google-cloud-secret-manager==2.18.4
google-cloud-aiplatform==1.38.1